output consistency: tidy up ignore class
separate abstract and mz-specific logic
