+++ misc/python/materialize/postgres_consistency/ignore_filter/pg_inconsistency_ignore_filter.py
@@ -27,13 +27,17 @@ from materialize.output_consistency.ignore_filter.expression_matchers import (
     matches_op_by_pattern,
     matches_x_or_y,
 )
-from materialize.output_consistency.ignore_filter.inconsistency_ignore_filter import (
+from materialize.output_consistency.ignore_filter.ignore_verdict import (
     IgnoreVerdict,
-    InconsistencyIgnoreFilter,
     NoIgnore,
+    YesIgnore,
+)
+from materialize.output_consistency.ignore_filter.inconsistency_ignore_filter import (
+    InconsistencyIgnoreFilter,
+)
+from materialize.output_consistency.ignore_filter.inconsistency_ignore_filter_base import (
     PostExecutionInconsistencyIgnoreFilterBase,
     PreExecutionInconsistencyIgnoreFilterBase,
-    YesIgnore,
 )
 from materialize.output_consistency.input_data.return_specs.date_time_return_spec import (
     DateTimeReturnTypeSpec,
